Move sub-services to specific directories
