+++ packages/app/services/registration/app-registrations.js
